Use CNI rather than host networking for buildkit (#478)
* Use CNI rather than host networking for buildkit

* A slightly more appropriate subnet address.

* Don't use CNI in earthly in earthly - it does not work due to need to share a CNI init lock.

* Minor.

* Kick all CIs.

+++ buildkitd/entrypoint.sh
@@ -20,6 +20,11 @@ if [ -z "$EARTHLY_TMP_DIR" ]; then
     exit 1
 fi
 
+if [ -z "$NETWORK_MODE" ]; then
+    echo "NETWORK_MODE not set"
+    exit 1
+fi
+
 if [ "$EARTHLY_RESET_TMP_DIR" = "true" ]; then
     echo "Resetting dir $EARTHLY_TMP_DIR"
     rm -rf "${EARTHLY_TMP_DIR:?}"/* || true
